Implement observer design pattern for search
I can't make this multithreaded I'm waiting on the search function to return, I need it to call a function when it's finished. Apparently the Observable / Observer interface stuff is deprecated but it looks like it was just done because it didn't provide enough functionality. But it does enough for what I need so I don't care. Also I was storing two lists of all the files so I removed one of them.